+++ src/clojureHover.ts
@@ -1,36 +1,31 @@
-'use strict';
-
-import * as fs from 'fs';
-import * as os from 'os';
-import * as path from 'path';
 import * as vscode from 'vscode';
-import {
-    nREPLClient
-} from './nreplClient';
-import {
-    ClojureProvider
-} from './clojureProvider';
 
-export class ClojureHoverProvider extends ClojureProvider implements vscode.HoverProvider {
+import { cljConnection } from './cljConnection';
+import { cljParser } from './cljParser';
+import { nreplClient } from './nreplClient';
 
-    provideHover(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): Thenable < vscode.Hover > {
-        return new Promise < vscode.Hover > ((resolve, reject) => {
-            let wordRange = document.getWordRangeAtPosition(position);
-            if (wordRange === undefined) {
-                resolve(new vscode.Hover('Docstring not found'));
-            } else {
-                let currentWord: string;
-                currentWord = document.lineAt(position.line).text.slice(wordRange.start.character, wordRange.end.character);
-                let ns = this.getNamespace(document.getText());
+export class ClojureHoverProvider implements vscode.HoverProvider {
 
-                let nrepl = this.getNREPL();
-                nrepl.info(currentWord, ns, (info) => {
-                    if (info.doc) {
-                        resolve(new vscode.Hover(info.doc));
-                    }
-                    reject();
-                });
-            }
+    provideHover(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): Thenable<vscode.Hover> {
+        if (!cljConnection.isConnected())
+            return Promise.reject('No nREPL connected.');
+
+        let wordRange = document.getWordRangeAtPosition(position);
+        if (wordRange === undefined)
+            return Promise.resolve(new vscode.Hover('Docstring not found'));
+
+        let currentWord: string;
+        currentWord = document.lineAt(position.line).text.slice(wordRange.start.character, wordRange.end.character);
+        const ns = cljParser.getNamespace(document.getText());
+
+        return cljConnection.sessionForFilename(document.fileName).then(session => {
+            return nreplClient.info(currentWord, ns, session.id).then(info => {
+                if (info.doc) {
+                    return Promise.resolve(new vscode.Hover(info.doc));
+                }
+                return Promise.reject(undefined);
+            });
         });
     }
+
 }