+++ src/clojureSuggest.ts
@@ -1,15 +1,7 @@
-'use strict';
-
-import * as fs from 'fs';
-import * as os from 'os';
-import * as path from 'path';
 import * as vscode from 'vscode';
-import {
-    nREPLClient
-} from './nreplClient';
-import {
-    ClojureProvider
-} from './clojureProvider';
+import { cljConnection } from './cljConnection';
+import { cljParser } from './cljParser';
+import { nreplClient } from './nreplClient';
 
 const mappings = {
     'nil': vscode.CompletionItemKind.Value,
@@ -20,58 +12,49 @@ const mappings = {
     'function': vscode.CompletionItemKind.Function,
     'special-form': vscode.CompletionItemKind.Keyword,
     'var': vscode.CompletionItemKind.Variable,
-    'method': vscode.CompletionItemKind.Method
+    'method': vscode.CompletionItemKind.Method,
 }
 
-export class ClojureCompletionItemProvider extends ClojureProvider implements vscode.CompletionItemProvider {
+export class ClojureCompletionItemProvider implements vscode.CompletionItemProvider {
 
-    public provideCompletionItems(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): Thenable < vscode.CompletionList > {
-        return new Promise < vscode.CompletionList > ((resolve, reject) => {
+    public provideCompletionItems(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): Thenable<vscode.CompletionList> {
+        if (!cljConnection.isConnected())
+            return Promise.reject('No nREPL connected.');
 
-            let document = vscode.window.activeTextEditor.document;
+        // TODO: Use VSCode means for getting a current word
+        let lineText = document.lineAt(position.line).text;
+        let words: string[] = lineText.split(' ');
+        let currentWord = words[words.length - 1].replace(/^[\('\[\{]+|[\)\]\}]+$/g, '');
+        let text = document.getText()
+        let ns = cljParser.getNamespace(text);
 
-            // TODO: Use VSCode means for getting a current word
-            let lineText = document.lineAt(position.line).text;
-            let words: string[] = lineText.split(' ');
-            let currentWord = words[words.length - 1].replace(/^[\('\[\{]+|[\)\]\}]+$/g, '');
-            let text = document.getText()
-            let ns = this.getNamespace(text);
+        let currentWordLength: number = currentWord.length;
 
-            let currentWordLength: number = currentWord.length;
+        let buildInsertText = (suggestion: string) => {
+            return suggestion[0] === '.' ? suggestion.slice(1) : suggestion;
+        }
 
-            let buildInsertText = (suggestion: string) => {
-                return suggestion[0] === '.' ? suggestion.slice(1) : suggestion;
-            }
+        return nreplClient.complete(currentWord, ns).then(completions => {
+            if (!('completions' in completions))
+                return Promise.reject(undefined);
 
-            let nrepl = this.getNREPL()
-            nrepl.complete(currentWord, ns, (completions) => {
-                let suggestions = [];
-                if ('completions' in completions) {
-                    completions.completions.forEach(element => {
-                        suggestions.push({
-                            label: element.candidate,
-                            kind: mappings[element.type] || vscode.CompletionItemKind.Text,
-                            insertText: buildInsertText(element.candidate)
-                        })
-                    })
-                } else {
-                    return reject();
-                }
-                let completionList: vscode.CompletionList = new vscode.CompletionList(suggestions, false);
-                resolve(completionList);
-
-            });
-        })
+            let suggestions = completions.completions.map(element => ({
+                label: element.candidate,
+                kind: mappings[element.type] || vscode.CompletionItemKind.Text,
+                insertText: buildInsertText(element.candidate)
+            }));
+            let completionList: vscode.CompletionList = new vscode.CompletionList(suggestions, false);
+            return Promise.resolve(completionList);
+        });
     }
 
-    public resolveCompletionItem(item: vscode.CompletionItem, token: vscode.CancellationToken): Thenable < vscode.CompletionItem > {
-        return new Promise < vscode.CompletionItem > ((resolve, reject) => {
-            let document = vscode.window.activeTextEditor.document;
-            let ns = this.getNamespace(document.getText());
-            this.getNREPL().info(item.label, ns, (info) => {
-                item.documentation = info.doc;
-                resolve(item);
-            })
-        })
+    public resolveCompletionItem(item: vscode.CompletionItem, token: vscode.CancellationToken): Thenable<vscode.CompletionItem> {
+        let document = vscode.window.activeTextEditor.document;
+        let ns = cljParser.getNamespace(document.getText());
+        return nreplClient.info(item.label, ns).then(info => {
+            item.documentation = info.doc;
+            return Promise.resolve(item);
+        });
     }
+
 }