+++ src/nreplClient.ts
@@ -1,9 +1,8 @@
-'use strict';
-
 import * as net from 'net';
 import { Buffer } from 'buffer';
 
 import * as bencodeUtil from './bencodeUtil';
+import { cljConnection, CljConnectionInformation } from './cljConnection';
 
 interface nREPLCompleteMessage {
     op: string;
@@ -15,12 +14,13 @@ interface nREPLInfoMessage {
     op: string;
     symbol: string;
     ns: string;
+    session: string;
 }
 
 interface nREPLEvalMessage {
     op: string;
     file: string;
-    'file-path'?: string,
+    'file-path'?: string;
     session: string;
 }
 
@@ -37,98 +37,112 @@ interface nREPLStacktraceMessage {
 
 interface nREPLCloneMessage {
     op: string;
+    session?: string;
 }
 
 interface nREPLCloseMessage {
     op: string;
+    session?: string;
 }
 
-export class nREPLClient {
+const complete = (symbol: string, ns: string): Promise<any> => {
+    const msg: nREPLCompleteMessage = { op: 'complete', symbol, ns };
+    return send(msg).then(respObjs => respObjs[0]);
+};
 
-    private host: string;
-    private port: number;
+const info = (symbol: string, ns: string, session?: string): Promise<any> => {
+    const msg: nREPLInfoMessage = { op: 'info', symbol, ns, session };
+    return send(msg).then(respObjs => respObjs[0]);
+};
 
-    public constructor(port: number, host: string) {
-        this.host = host;
-        this.port = port;
-    }
+const evaluate = (code: string, session?: string): Promise<any[]> => clone(session).then((new_session) => {
+    const session_id = new_session['new-session'];
+    const msg: nREPLSingleEvalMessage = { op: 'eval', code: code, session: session_id };
+    return send(msg);
+});
 
-    public complete(symbol: string, ns: string, callback) {
-        const msg: nREPLCompleteMessage = { op: 'complete', symbol: symbol, ns: ns };
-        this.send(msg).then(respObjs => callback(respObjs[0]));
-    }
+const evaluateFile = (code: string, filepath: string, session?: string): Promise<any[]> => clone(session).then((new_session) => {
+    const session_id = new_session['new-session'];
+    const msg: nREPLEvalMessage = { op: 'load-file', file: code, 'file-path': filepath, session: session_id };
+    return send(msg);
+});
 
-    public info(symbol: string, ns: string, callback) {
-        const msg: nREPLInfoMessage = { op: 'info', symbol: symbol, ns: ns };
-        this.send(msg).then(respObjs => callback(respObjs[0]));
-    }
+const stacktrace = (session: string): Promise<any> => send({ op: 'stacktrace', session: session });
 
-    public eval(code: string): Promise<any[]> {
-        return this.clone().then((new_session) => {
-            const session_id = new_session['new-session'];
-            const msg: nREPLSingleEvalMessage = { op: 'eval', code: code, session: session_id };
-            return this.send(msg);
+const clone = (session?: string): Promise<any[]> => send({ op: 'clone', session: session }).then(respObjs => respObjs[0]);
+
+const test = (connectionInfo: CljConnectionInformation): Promise<any[]> => {
+    return send({ op: 'clone' }, connectionInfo)
+        .then(respObjs => respObjs[0])
+        .then(response => {
+            if (!('new-session' in response))
+                return Promise.reject(false);
         });
-    }
+};
 
-    public evalFile(code: string, filepath: string): Promise<any[]> {
-        return this.clone().then((new_session) => {
-            const session_id = new_session['new-session'];
-            const msg: nREPLEvalMessage = { op: 'load-file', file: code, 'file-path': filepath, session: session_id };
-            return this.send(msg);
+const close = (session?: string): Promise<any[]> => send({ op: 'close', session: session });
+
+const listSessions = (): Promise<[string]> => {
+    return send({op: 'ls-sessions'}).then(respObjs => {
+        const response = respObjs[0];
+        if (response.status[0] == "done") {
+            return Promise.resolve(response.sessions);
+        }
+    });
+}
+
+const send = (msg: nREPLCompleteMessage | nREPLInfoMessage | nREPLEvalMessage | nREPLStacktraceMessage | nREPLCloneMessage | nREPLCloseMessage | nREPLSingleEvalMessage, connection?: CljConnectionInformation): Promise<any[]> => {
+    return new Promise<any[]>((resolve, reject) => {
+        connection = connection || cljConnection.getConnection();
+
+        if (!connection)
+            return reject('No connection found.');
+
+        const client = net.createConnection(connection.port, connection.host);
+        Object.keys(msg).forEach(key => msg[key] === undefined && delete msg[key]);
+        client.write(bencodeUtil.encode(msg), 'binary');
+
+        client.on('error', error => {
+            client.end();
+            client.removeAllListeners();
+            reject(error);
         });
-    }
 
-    public stacktrace(session: string): Promise<any> {
-        const msg: nREPLStacktraceMessage = { op: 'stacktrace', session: session };
-        return this.send(msg);
-    }
+        let nreplResp = Buffer.from('');
+        const respObjects = [];
+        client.on('data', data => {
+            nreplResp = Buffer.concat([nreplResp, data]);
+            const { decodedObjects, rest } = bencodeUtil.decodeObjects(nreplResp);
+            nreplResp = rest;
+            const validDecodedObjects = decodedObjects.reduce((objs, obj) => {
+                if (!isLastNreplObject(objs))
+                    objs.push(obj);
+                return objs;
+            }, []);
+            respObjects.push(...validDecodedObjects);
 
-    public clone(): Promise<any[]> {
-        const msg = { op: 'clone' };
-        return this.send(msg).then(respObjs => respObjs[0]);
-    }
-
-    public close(): Promise<any[]> {
-        const msg: nREPLCloseMessage = { op: 'close' };
-        return this.send(msg);
-    }
-
-    private send(msg: nREPLCompleteMessage | nREPLInfoMessage | nREPLEvalMessage | nREPLStacktraceMessage | nREPLCloneMessage | nREPLCloseMessage | nREPLSingleEvalMessage): Promise<any[]> {
-        return new Promise<any[]>((resolve, reject) => {
-            const client = net.createConnection(this.port, this.host);
-            client.write(bencodeUtil.encode(msg), 'binary');
-
-            client.on('error', error => {
+            if (isLastNreplObject(respObjects)) {
                 client.end();
                 client.removeAllListeners();
-                reject(error);
-            });
-
-            let nreplResp = Buffer.from('');
-            const respObjects = [];
-            client.on('data', data => {
-                nreplResp = Buffer.concat([nreplResp, data]);
-                const { decodedObjects, rest } = bencodeUtil.decodeObjects(nreplResp);
-                nreplResp = rest;
-                const validDecodedObjects = decodedObjects.reduce((objs, obj) => {
-                    if (!isLastNreplObject(objs))
-                        objs.push(obj);
-                    return objs;
-                }, []);
-                respObjects.push(...validDecodedObjects);
-
-                if (isLastNreplObject(respObjects)) {
-                    client.end();
-                    client.removeAllListeners();
-                    resolve(respObjects);
-                }
-            });
+                resolve(respObjects);
+            }
         });
-    }
-}
+    });
+};
 
-function isLastNreplObject(nreplObjects: any[]): boolean {
+const isLastNreplObject = (nreplObjects: any[]): boolean => {
     const lastObj = [...nreplObjects].pop();
     return lastObj && lastObj.status && lastObj.status.indexOf('done') > -1;
 }
+
+export const nreplClient = {
+    complete,
+    info,
+    evaluate,
+    evaluateFile,
+    stacktrace,
+    clone,
+    test,
+    close,
+    listSessions
+};