+++ CHANGELOG.md
@@ -1,3 +1,68 @@
+# Version 0.9.1
+
+Adds a configuration option for formatting code on save.
+
+# Version 0.9.0
+
+Adds experimental ClojureScript support. Please check out [README.md](https://github.com/avli/clojureVSCode#clojurescript-project-setup) to learn how to use the extension for ClojureScript.
+
+# Version 0.8.2
+
+Fixes the issue with unescaping newlines that are parts of a string [64](https://github.com/avli/clojureVSCode/issues/64).
+
+# Version 0.8.1
+
+Adds a configuration option to disable an embedded nREPL start.
+
+# Version 0.8.0
+
+Adds code formatting support [57](https://github.com/avli/clojureVSCode/issues/57).
+
+# Version 0.7.10
+
+Adds workaround for the hanging Java process issue on the Windows platform [56](https://github.com/avli/clojureVSCode/issues/56).
+
+# Version 0.7.9
+
+Hotfix that detaches nREPL process on UNIX and non-detached one on Windows [56](https://github.com/avli/clojureVSCode/issues/56).
+
+# Version 0.7.8
+
+Fixes the troubles with a hanging Java process [56](https://github.com/avli/clojureVSCode/issues/56).
+
+# Version 0.7.7
+
+Fixes the error with embedded nREPL on Windows.
+
+# Version 0.7.6
+
+Does code refactoring and minor nREPL connection indicator UI fixes.
+
+# Version 0.7.5
+
+Fixes the `Unable to start nREPL` bug [47](https://github.com/avli/clojureVSCode/issues/47).
+
+# Version 0.7.4
+
+Fixes the nREPL connection indicator behavior [42](https://github.com/avli/clojureVSCode/issues/42).
+
+# Version 0.7.3
+
+Fixes typos in documentation.
+
+# Version 0.7.2
+
+Fixes README.md formatting to display correctly on the Visual Studio Marketplace web-site.
+
+# Version 0.7.1
+
+Fixes the links to the images in README.md.
+
+# Version 0.7.0
+
+Adds embedded nREPL functionality – no need to connect manually anymore.
+[21](https://github.com/avli/clojureVSCode/issues/21).
+
 # Version 0.6.1
 
 Adds the forgotten changelog for the version 0.6.0 :-)